% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{AflatoxinLiverCancer_df}
\alias{AflatoxinLiverCancer_df}
\title{Aflatoxin Dosage and Liver Cancer in Lab Animals}
\format{
A data frame with 6 observations and 3 variables:
\describe{
  \item{dose}{Dose of Aflatoxin B1 administered (integer).}
  \item{total}{Total number of animals exposed to the dose (integer).}
  \item{tumor}{Number of animals that developed liver cancer (integer).}
}
}
\source{
Data taken from the faraway package.
Gaylor DW (1987). *Linear nonparametric upper limits for low dose extrapolation*. ASA Proceedings of the Biopharmaceutical Section.
}
\usage{
data(AflatoxinLiverCancer_df)
}
\description{
This dataset, AflatoxinLiverCancer_df, is a data frame containing data from a study where varying doses of Aflatoxin B1 were administered to lab animals. The dataset records the total number of animals exposed to each dose and the number of animals that developed liver cancer.
}
\details{
The dataset name has been kept as 'AflatoxinLiverCancer_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying
its specific characteristics. The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.
}
