% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{AlcoholIntakeCancer_df}
\alias{AlcoholIntakeCancer_df}
\title{Alcohol Intake and Colorectal Cancer Data}
\format{
A data frame with 48 observations and 7 variables:
\describe{
  \item{id}{Identifier for the study (factor).}
  \item{type}{Type of study (factor).}
  \item{dose}{Level of alcohol intake (numeric).}
  \item{cases}{Number of colorectal cancer cases (integer).}
  \item{peryears}{Person-years of observation (numeric).}
  \item{logrr}{Logarithm of the relative risk (numeric).}
  \item{se}{Standard error of the logarithm of the relative risk (numeric).}
}
}
\source{
Data taken from the mixmeta package. Available at: https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041
}
\usage{
data(AlcoholIntakeCancer_df)
}
\description{
This dataset, AlcoholIntakeCancer_df, is a data frame containing data related to alcohol intake and its association with colorectal cancer risk.
The data includes information on alcohol intake levels (dose), the number of cancer cases, person-years of observation, and the relative risk (logrr) along with its standard error (se).
The dataset consists of 48 observations with 7 variables.
}
\details{
The dataset name has been kept as 'AlcoholIntakeCancer_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying
its specific characteristics. The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.
}
