% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{BRCA2BreastCancer_df}
\alias{BRCA2BreastCancer_df}
\title{Cumulative Risk of Women Breast Cancer BRCA2 Mutation}
\format{
A data frame with 11 observations and 2 variables:
\describe{
  \item{x}{Age of the individual in years (numeric).}
  \item{y}{Cumulative risk of breast cancer at that age (numeric).}
}
}
\source{
Data taken from the riskyr package.
}
\usage{
data(BRCA2BreastCancer_df)
}
\description{
This dataset, BRCA2BreastCancer_df, is a data frame containing data on the cumulative risk of breast cancer in women with the BRCA2 mutation as a function of their age.
The dataset includes 11 observations, with each entry representing the cumulative risk at a specific age (in years).
}
\details{
The dataset name has been kept as 'BRCA2BreastCancer_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying
its specific characteristics. The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.
}
