% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{BRCA2OvarianCancer_df}
\alias{BRCA2OvarianCancer_df}
\title{Cumulative Risk of Women Ovarian Cancer BRCA2 Mutation}
\format{
A data frame with 63 observations and 2 variables:
\describe{
  \item{age}{Age of the individual in years (numeric).}
  \item{cumRisk}{Cumulative risk of ovarian cancer at that age (numeric).}
}
}
\source{
Data taken from the riskyr package.
Based on Figure 2 (p. 2408) of Kuchenbaecker, K. B., Hopper, J. L., Barnes, D. R., Phillips, K. A., Mooij, T. M., Roos-Blom, M. J., ... & BRCA1 and BRCA2 Cohort Consortium (2017). Risks of breast, ovarian, and contralateral breast cancer for BRCA1 and BRCA2 mutation carriers. JAMA, 317 (23), 2402--2416. doi: 10.1001/jama.2017.7112
}
\usage{
data(BRCA2OvarianCancer_df)
}
\description{
This dataset, BRCA2OvarianCancer_df, is a data frame containing data on the cumulative risk of ovarian cancer in women with the BRCA2 mutation as a function of their age.
The dataset includes 63 observations, with each entry representing the cumulative risk at a specific age (in years).
}
\details{
The dataset name has been kept as 'BRCA2OvarianCancer_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying
its specific characteristics. The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.
}
