% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{LeukemiaLymphomaControl_df}
\alias{LeukemiaLymphomaControl_df}
\title{North Humberside Leukemia and Lymphoma Control Cases}
\format{
A data frame with 191 observations and 2 variables:
\describe{
  \item{locationid}{Location ID (integer).}
  \item{numcontrols}{Number of control cases (integer).}
}
}
\source{
Data taken from the rsatscan package, distributed with SaTScan software: https://www.satscan.org
}
\usage{
data(LeukemiaLymphomaControl_df)
}
\description{
This dataset, LeukemiaLymphomaControl_df, is a data frame containing information on the number of control cases
for leukemia and lymphoma reported in different locations within North Humberside. The dataset includes the location ID and the number of control cases for each location.
}
\details{
The dataset name has been kept as 'LeukemiaLymphomaControl_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying
its specific characteristics. The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.
}
