% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{LeukemiaLymphomaGeo_df}
\alias{LeukemiaLymphomaGeo_df}
\title{North Humberside Leukemia and Lymphoma Geographic Data}
\format{
A data frame with 191 observations and 3 variables:
\describe{
  \item{locationid}{Location ID (integer).}
  \item{x-coordinate}{X-coordinate (integer).}
  \item{y-coordinate}{Y-coordinate (integer).}
}
}
\source{
Data taken from the rsatscan package, distributed with SaTScan software: https://www.satscan.org
}
\usage{
data(LeukemiaLymphomaGeo_df)
}
\description{
This dataset, LeukemiaLymphomaGeo_df, is a data frame containing the geographical coordinates (x and y)
for locations in North Humberside related to leukemia and lymphoma cases. It includes the location ID and the coordinates
for each of the 191 locations.
}
\details{
The dataset name has been kept as 'LeukemiaLymphomaGeo_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying
its specific characteristics. The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.
}
