% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{LeukemiaSurvival_df}
\alias{LeukemiaSurvival_df}
\title{Leukemia Remission Survival Times Placebo-Controlled RCT}
\format{
A data frame with 42 observations and 5 variables:
\describe{
  \item{time}{Time to remission in days (integer).}
  \item{status}{Patient status (1 for event, 0 for censored) (integer).}
  \item{sex}{Gender of the patient (numeric, 1 for male, 2 for female).}
  \item{logWBC}{Log-transformed white blood cell count (numeric).}
  \item{rx}{Treatment regimen (numeric, coded treatment type).}
}
}
\source{
Data taken from the autoReg package.
}
\usage{
data(LeukemiaSurvival_df)
}
\description{
This dataset, LeukemiaSurvival_df, is a data frame containing remission survival times of 42 leukemia patients enrolled in a placebo-controlled randomized controlled trial (RCT).
The dataset includes information on the time to remission, patient status, sex, white blood cell count (log-transformed), and treatment regimen.
}
\details{
The dataset name has been kept as 'LeukemiaSurvival_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying
its specific characteristics. The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.
}
