% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncodatasets-package.R
\docType{package}
\name{OncoDataSets}
\alias{OncoDataSets}
\alias{OncoDataSets-package}
\title{OncoDataSets: A Comprehensive Collection of Cancer Types and Cancer-related DataSets}
\description{
{
This package provides a wide variety of datasets related to cancer types such as melanoma, leukemia, breast, ovarian, and lung cancer, among others.
}
}
\details{
OncoDataSets: A Comprehensive Collection of Cancer Types and Cancer-related DataSets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Comprehensive Collection of Cancer Types and Cancer-related DataSets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/oncodatasets}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
