% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{SmokingLungCancer_df}
\alias{SmokingLungCancer_df}
\title{Years of Smoking and Lung Cancer Deaths in Men}
\format{
A data frame with 63 observations and 4 variables:
\describe{
  \item{yrs_smk}{Years of smoking, represented as a factor with 9 levels.}
  \item{pys}{Pack-years of smoking (numeric).}
  \item{num_cigs}{Number of cigarettes smoked per day, represented as a factor with 7 levels.}
  \item{deaths}{Number of deaths due to lung cancer (numeric).}
}
}
\source{
Data taken from the R4HCR package.
Data originally from Table 24-4, page 702 of Kleinbaum et al (1988).
}
\usage{
data(SmokingLungCancer_df)
}
\description{
This dataset, SmokingLungCancer_df, is a data frame containing data on man-years of risk and observed number of lung cancer deaths among men.
The data includes information about the years of smoking, pack-years, number of cigarettes smoked per day, and the number of deaths due to lung cancer.
}
\details{
The dataset name has been kept as 'SmokingLungCancer_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying
its specific characteristics. The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.
}
