% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{USCancerStats_df}
\alias{USCancerStats_df}
\title{US Cancer Incidence, Mortality, and Survival Changes}
\format{
A data frame with 20 observations and 4 variables:
\describe{
  \item{site}{Tumor site (character).}
  \item{survival}{Absolute difference in 5-year survival (numeric).}
  \item{mortality}{Percentage change in mortality (numeric).}
  \item{incidence}{Percentage change in incidence (numeric).}
}
}
\source{
Data taken from the R4HCR package.
}
\usage{
data(USCancerStats_df)
}
\description{
This dataset, USCancerStats_df, is a data frame containing cancer statistics for 20 solid tumor types, including
incidence, mortality, and survival data. The dataset reports the absolute difference in 5-year survival between 1989-1995 and 1950-1954,
as well as the percentage change in mortality and incidence from 1950 to 1996.
}
\details{
The dataset name has been kept as 'USCancerStats_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying
its specific characteristics. The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.
}
