% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{VALungCancer_list}
\alias{VALungCancer_list}
\title{VA Lung Cancer Data Set}
\format{
A list with 2 components:
\describe{
  \item{X}{A numeric matrix with 1137 rows and 19 columns, representing the covariates.}
  \item{y}{A numeric matrix with 1137 rows and 12 columns, representing the survival time data. The columns include 'time' for the survival time and other variables related to survival analysis.}
}
}
\source{
Data taken from the ncvreg package.
Based on data from a randomized trial of two treatment regimens for lung cancer, as presented in the classic textbook by Kalbfleisch and Prentice.
}
\usage{
data(VALungCancer_list)
}
\description{
This dataset, VALungCancer_list, is a list containing two components: `X` and `y`. The data comes from a randomized trial of two treatment regimens for lung cancer. The `X` component contains the covariates, and the `y` component contains the survival time data. This dataset is typically used in survival analysis.
}
\details{
The dataset name has been kept as 'VALungCancer_list' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying
its specific characteristics. The suffix '_list' indicates that the dataset is a list. The original content has not been modified in any way.
}
